requirements.txt
html2text
beautifulsoup4
trafilatura
httpx
bleach
python-fasthtml

